Pass long longs to the testL function and use the ll modifier to print them with printf. Bug found by Vladimir Prus
llvm-svn: 14688

+++ SingleSource/UnitTests/2002-05-19-DivTest.c
@@ -1,8 +1,8 @@
 extern int printf(const char *, ...);
 
-void testL(long Arg) {
-  printf("%ld\n", Arg / (1LL << 4));
-  printf("%ld\n", Arg / (1LL << 46));
+void testL(long long Arg) {
+  printf("%lld\n", Arg / (1LL << 4));
+  printf("%lld\n", Arg / (1LL << 46));
 }
 
 void test(int Arg) {
